5398: move dateinput and dateinputs
